Extract decideWatcherBinding so unit tests avoid DATABASE_URL.
